refactor(tests): consolidate test harness, relocate generic related helpers
God-node analysis remediation (docs/god-nodes-analysis.md):

- Move generic relatedResult + assertStruct out of internal/aws/ec2_related.go
  into new internal/aws/related_common.go (same package aws, zero call-site
  change) — they serve every *_related.go checker, not just EC2.
- Add tests/unit/tuitest package as the single canonical harness kit
  (Step, StepModel, Render, Sized, StripANSI, Width, NoColor); reduce 9
  helpers duplicated across the unit / unit_test package boundary to thin
  forwarders. Names unchanged, so ~150 call sites are untouched. Removes the
  cross-package ansiRegex duplicate.
- Refresh stale CLAUDE.md glob counts surfaced during the analysis
  (tests/unit 148->710, internal/aws 77->357).

Behavior unchanged. make build / make test / make lint all green.
