+++ internal/aws/ec2_related.go
@@ -19,19 +19,6 @@ import (
 	"github.com/k2m30/a9s/v3/internal/resource"
 )
 
-// assertStruct extracts a value of type T from an interface that may hold
-// either T or *T. Used for RawStruct type assertions across related checkers.
-func assertStruct[T any](v any) (T, bool) {
-	if val, ok := v.(T); ok {
-		return val, true
-	}
-	if p, ok := v.(*T); ok && p != nil {
-		return *p, true
-	}
-	var zero T
-	return zero, false
-}
-
 // checkEC2TargetGroups checks the cache for target groups referencing this EC2 instance.
 func checkEC2TargetGroups(ctx context.Context, clients any, res resource.Resource, cache resource.ResourceCache) resource.RelatedCheckResult {
 	if res.RawStruct == nil {
@@ -377,30 +364,6 @@ func ec2RelatedResources(ctx context.Context, clients any, cache resource.Resour
 		}
 	}
 	return resources, isTruncated, err
-}
-
-func relatedResult(target string, ids []string) resource.RelatedCheckResult {
-	if len(ids) == 0 {
-		return resource.RelatedCheckResult{TargetType: target, Count: 0}
-	}
-	set := make(map[string]struct{}, len(ids))
-	uniq := make([]string, 0, len(ids))
-	for _, id := range ids {
-		if id == "" {
-			continue
-		}
-		if _, ok := set[id]; ok {
-			continue
-		}
-		set[id] = struct{}{}
-		uniq = append(uniq, id)
-	}
-	sort.Strings(uniq)
-	return resource.RelatedCheckResult{
-		TargetType:  target,
-		Count:       len(uniq),
-		ResourceIDs: uniq,
-	}
 }
 
 func ec2Identity(res resource.Resource) (instanceID, vpcID, stackName string) {
