+++ internal/aws/related_common.go
@@ -1,0 +1,45 @@
+// Generic helpers shared by every *_related.go checker (moved out of ec2_related.go — they are not EC2-specific).
+package aws
+
+import (
+	"sort"
+
+	"github.com/k2m30/a9s/v3/internal/resource"
+)
+
+// assertStruct extracts a value of type T from an interface that may hold
+// either T or *T. Used for RawStruct type assertions across related checkers.
+func assertStruct[T any](v any) (T, bool) {
+	if val, ok := v.(T); ok {
+		return val, true
+	}
+	if p, ok := v.(*T); ok && p != nil {
+		return *p, true
+	}
+	var zero T
+	return zero, false
+}
+
+func relatedResult(target string, ids []string) resource.RelatedCheckResult {
+	if len(ids) == 0 {
+		return resource.RelatedCheckResult{TargetType: target, Count: 0}
+	}
+	set := make(map[string]struct{}, len(ids))
+	uniq := make([]string, 0, len(ids))
+	for _, id := range ids {
+		if id == "" {
+			continue
+		}
+		if _, ok := set[id]; ok {
+			continue
+		}
+		set[id] = struct{}{}
+		uniq = append(uniq, id)
+	}
+	sort.Strings(uniq)
+	return resource.RelatedCheckResult{
+		TargetType:  target,
+		Count:       len(uniq),
+		ResourceIDs: uniq,
+	}
+}
