+++ tests/unit/helpers_test.go
@@ -3,17 +3,18 @@ package unit
 import (
 	"regexp"
 
-	lipgloss "charm.land/lipgloss/v2"
+	"github.com/k2m30/a9s/v3/tests/unit/tuitest"
 )
 
+// ansiRe is kept for direct call sites in search tests that use it inline.
 var ansiRe = regexp.MustCompile(`\x1b\[[0-9;]*[a-zA-Z]`)
 
 // stripANSI removes ANSI escape sequences from a string for plain-text comparison.
 func stripANSI(s string) string {
-	return ansiRe.ReplaceAllString(s, "")
+	return tuitest.StripANSI(s)
 }
 
 // lipglossWidth measures the visible width of a string (ANSI-aware).
 func lipglossWidth(s string) int {
-	return lipgloss.Width(s)
+	return tuitest.Width(s)
 }
